+++ server/utils/database.js
@@ -3,35 +3,7 @@ const bcrypt = require("bcrypt");
 
 const { DB_USER, DB_PASS, DB_NAME, DB_TEST_NAME } = process.env;
 const saltRounds = 10;
-const mentionSchema = new mongoose.Schema({
-  content: {
-    type: String,
-    
-  },
-  title: {
-    type: String,
-    required: true,
-  },
-  platform: {
-    type: String,
-    required: true,
-  },
-  image: {
-    type: String,
-  },
-  date: {
-    type: String,
-    required: true,
-  },
-  popularity: {
-    type: String,
-  },
-  url: {
-    type:String,
-  }
 
-});
-const mention = mongoose.model('mention', mentionSchema);
 /**
  * Opens a connection to the MongoDB Atlas instance used, using the
  * individual login credentials created in the root environment file.
@@ -48,19 +20,6 @@ const connectDB = (environment = "prod") => {
   mongoose.connect(URI, options);
 };
  
-function createMention(mentionBody, mentionTitle, mentionPlatform, mentionImage, mentionDate, mentionPopularity, mentionUrl){
-  const newMention = new mention({content: mentionBody, title: mentionTitle, platform: mentionPlatform, image: mentionImage,date: mentionDate, popularity:mentionPopularity, url: mentionUrl});
-  newMention.save(function (err) { if (err) return console.error(err);});
-}
-function getMention(){
-  let ans =[];
-  ans.push(mention.find(function (err, mention) {
-  if (err) return console.error(err);
-    return mention;
-  }))
-  
-  return ans;
-}
 
 /**
  * Closes the default mongoose connection.
@@ -105,7 +64,6 @@ module.exports = {
   connectDB,
   disconnectDB,
   databaseErrorHandler,
-  encryptPasswordWithSalt,
-  createMention,
-  getMention
+  encryptPasswordWithSalt
+
 };
